services/device-hub/requirements.txt

paho-mqtt==2.1.0
psycopg[binary]==3.2.1
python-dotenv==1.0.1
